+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -84,10 +84,21 @@ class SupportProcessorTestCase(KBBioProcessTestCase):
             expression_3 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")
             expression_4 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")
 
+            multiqc = self.run_process('multiqc', {
+                'data': [
+                    bam.id,
+                    reads.id,
+                ],
+                'advanced': {
+                    'dirs': True,
+                    'config': True,
+                }
+            })
+
         self.run_process('archive-samples', {
             'data': [txt_file.id, bam.id, reads.id, vcf.id, expression_1.id, expression_2.id, expression_5.id,
-                     expression_4.id, expression_3.id],
-            'fields': ['file', 'bam', 'bai', 'fastq', 'fastqc_url', 'fastqc_archive', 'vcf', 'exp']})
+                     expression_4.id, expression_3.id, multiqc.id],
+            'fields': ['file', 'bam', 'bai', 'fastq', 'fastqc_url', 'fastqc_archive', 'vcf', 'exp_set', 'report']})
 
     @tag_process('prepare-geo-chipseq')
     def test_prepare_geo_chipseq(self):